Fix PointCloud2Iterator type mismatch on per-point time field
The `time` field is declared FLOAT64 in setPointCloud2Fields() but
written through a PointCloud2Iterator<float>. Each iteration writes
only the low 4 bytes of the 8-byte slot, leaving the upper 4 bytes at
the buffer's zero-init value. Consumers that read the field as double
(per the message declaration) get denormal values around 1e-300 in
little-endian, instead of the per-point relative timestamp.

The v0.2.0 changelog notes "Changed point time from float to double —
to maintain time precision downstream", but the iterator template
parameter wasn't widened along with the field declaration, so the
upgrade only half-landed. This fix completes it.

Affected consumers: motion-compensation / deskewing pipelines that
read the time field (LIO-SAM, FAST-LIO, Cartographer with motion
comp). Standard PCL filters, rviz, and Nav2 obstacle layers are
unaffected — they don't read `time`.

+++ src/l2lidar_node.cpp
@@ -412,7 +412,9 @@ void L2LidarNode::onPointCloudReceived()
     sensor_msgs::PointCloud2Iterator<float> iter_z(cloud, "z");
     sensor_msgs::PointCloud2Iterator<float> iter_i(cloud, "intensity");
     sensor_msgs::PointCloud2Iterator<float> iter_r(cloud, "range");
-    sensor_msgs::PointCloud2Iterator<float> iter_t(cloud, "time");
+    // Field declared FLOAT64 above; iterator must match width or upper
+    // 4 bytes of every 8-byte slot stay uninitialized.
+    sensor_msgs::PointCloud2Iterator<double> iter_t(cloud, "time");
 
     if(UseAggFrame) {
         modifier.resize(aggframe.size());
